ghost_app_cdk/infra_cdk_code/requirements.txt
ruamel_yaml
boto3

aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
